Don't rely on load order for built-in classes.
Generate these classes using the structures that the runtime expects
internally, rather than relying on the Objective-C compiler.  This
change means that they can always be the latest version, even if the
runtime is compiled with an older compiler, and ensures that the
`Protocol` class is always available, independent of global constructor
ordering between libraries.

Fixes #283